refactor: move hdfs specific logic and test to storage_hdfs
